Created Person-IDs for donors of BMLacq811 (DamaKrestos and Mikael)
The note in BMLacq811 (according to Marrassini) says: ዝመጽሐፍ፡ ዘዘ<persName ref="PRS13981Mikael">ሚካኤል፡</persName>
ወዘአቡሁ፡ <persName ref="PRS13980DamaKrestos">ደመ፡ ክርስቶስ፡</persName> ዘሰረቆ፡ ወዘፈሐቆ፡ ውጉዘ፡ ለይኩን፡
I understand, that Dama Krestos is the father of Mikael and together they donated this book to receive pardon for whatever they did. I selected "DonationNote" in the description of BMLacq811, but in the person document I could not find donation or donor and selected "Owner" for that reason. Is that correct?