Created Person-IDs for donors of BMLacq811 (DamaKrestos and Mikael) (#687)
* Created Person-IDs for donors of BMLacq811 (DamaKrestos and Mikael)

The note in BMLacq811 (according to Marrassini) says: ዝመጽሐፍ፡ ዘዘ<persName ref="PRS13981Mikael">ሚካኤል፡</persName>
ወዘአቡሁ፡ <persName ref="PRS13980DamaKrestos">ደመ፡ ክርስቶስ፡</persName> ዘሰረቆ፡ ወዘፈሐቆ፡ ውጉዘ፡ ለይኩን፡
I understand, that Dama Krestos is the father of Mikael and together they donated this book to receive pardon for whatever they did. I selected "DonationNote" in the description of BMLacq811, but in the person document I could not find donation or donor and selected "Owner" for that reason. Is that correct?

* Create PRS13982Estifanos.xml

Scribe of BMLacq811 as attested in the manuscript (following Marrassini)

* Update PRS13982Estifanos.xml

* Update relations PRS13890 DamaKrestos and PRS13891Mikael

* Change name to Zamikael in PRS13981Mikael to PRS13981Zamikael

Changed name and title of Mikael zu Zamikael

* Update PRS13981Mikael.xml

I restored ID PRS13981Mikael for the person Zamikāʾel, son of Dama Krəstos, mentioned in BMLacq811. I wonder why should not create a new ID, since now the PRS-ID does not correspond to the name any more.

* Updated relations in PRS13981Mikael and PRS13980DamaKrestos

I corrected the xml:ID for PRS13981Mikael yesterday, but probably did not push it orderly. I hope it will now be everything correct.

* Update PRS13981Mikael.xml

title fix

* Marked up floruit in three PRS-ID from BMLacq811

I marked up floruit in PRS13980DamaKrestos, PRS13981Mikael and 13982Estifanos and switched translitteration and title.

---------

Co-authored-by: Eugenia <eugenia.sokolinski@uni-hamburg.de>