Create PRS13990ZawaldaMaryam.xml
The name Zawalda Maryam occurs in some of the prayers rather in the middle of the manuscript for example on f. 12rb. Several persons with the name Zawalda Māryām and a period of activity in the 19th century already exist as for instance PRS11836ZawaldaM or PRS12182ZawaldaM, but I think it would be rather accidental if the owner of this manuscript was indeed one of them and we have no chance to make that certain. For this reason I rather created a new ID.