Create PRS13990ZawaldaMaryam.xml (#693)
* Create PRS13990ZawaldaMaryam.xml

The name Zawalda Maryam occurs in some of the prayers rather in the middle of the manuscript for example on f. 12rb. Several persons with the name Zawalda Māryām and a period of activity in the 19th century already exist as for instance PRS11836ZawaldaM or PRS12182ZawaldaM, but I think it would be rather accidental if the owner of this manuscript was indeed one of them and we have no chance to make that certain. For this reason I rather created a new ID.

* Updated Zawalda Maryam

Updated title and floruit according to comments of Doro and Denis.

* Identification of Scribe and Owner Zawalda Maryam

Denis has informed me, that the Scribe and Owner of BMLacq810 is identical with the scribe of BMLacq306 and various other manuscripts originating from Addigrat. A new PRS-ID is not necessary any more. The ID PRS13990ZawaldaMaryam, that I created yesterday can be deleted or dumped.