New PRS-ID for GabraHanaCare, KidanaMaryam and TaklaMaryam
I created some new ID for persons that are to be found in BMLacq720. The name Gabra Ḥanā Čare is not attested before in betamaṣaḥəft. The name Kidāna Māryām on the other hand is present a few times, but the relation is to unspecific to identify him or her with an existant PRS-ID. The scribe Takla Māryām is perhaps the same as the one, who wrote EMIP03224 - at least the name is identical, but alas I cannot check, whether it is the same hand, because there are not images available. I do not even have a clue, whether the possible date may allow such a statement, because this is not encoded in the record of EMIP03224.