Created IDs PRS13997Valeriani and PRS13998Formigli (#700)
* Created IDs PRS13997Valeriani and PRS13998Formigli

D. Valeriani wrote in a letter to Formigli informing him about the contents of BMLacq38. This paper letter is attached to BMLacq38. I cannot read the old Italian writing and Marrassini is very short in summarizing the content. This is why I cannot say much about these gentlemen.

* Updated PRS13997Valeriani and PRS13998Formigli

According to Eugenias comments

* Updated PRS13997Valeriani and PRS13998Formigli

* Update PRS13998Formigli