Updated PRS in BMLacq680
I updated various PRS IDs for persons that are attested in BMLacq680.
First: I filled the empty record of PRS14025Balca with content. This content is taken from PRS2435Balcasa (I changed just the title in the title statement and in the header). The latter was reworked by myself recently. The first is to be deleted. The other shall remain.
I changed the markUp of dates in PRS14031John according to Eugenias request.
I changed the title in a relation in PRS2335Balcasa that I happened to mention incidantally.
I created PRS14034FalibulaKayna as discussed in Issue #2365. I have no english names for this couple of saints and no additional information. I can imagine, that the name Fālibulā is coming from Philippa or Kāynā from Cain, as in the Bible. But this mere speculation and I do not want to put any english or greek names into the record, just yet alone the latin translitteration from Dillmanns catalogue and the diverging Gəʿəz names. I don't know, how to refer to the Bodleian manuscript. There is no number in Dillmanns catalogue except for his own (XXII), but I don't find BDLaeth with the number 22.