Updated and created PRS IDs related to BMLacq680 (#720)
* Updated and created PRS IDs related to BMLacq680

Updated PRS IDs related to BMLacq680 and created new ID for Bālčā Śāfo being the former owner of the manuscript.

* Update PRS14025Balca

According to Theas quests.

* Updated PRS2435Balcasa and created PRS14031

I created a new ID for Yoḥannəs of Burləs and updated the PRS ID of Daǧǧazmač Balča Śāfo. In turn the newly created ID (PRS14025Balca) can be deleted after merging.

* Updated PRS14031John

Updated relation Authorship Synaxary and estimated dates of birth and death.

* Updated PRS in BMLacq680

I updated various PRS IDs for persons that are attested in BMLacq680.
First: I filled the empty record of PRS14025Balca with content. This content is taken from PRS2435Balcasa (I changed just the title in the title statement and in the header). The latter was reworked by myself recently. The first is to be deleted. The other shall remain.
I changed the markUp of dates in PRS14031John according to Eugenias request.
I changed the title in a relation in PRS2335Balcasa that I happened to mention incidantally.
I created PRS14034FalibulaKayna as discussed in Issue #2365. I have no english names for this couple of saints and no additional information. I can imagine, that the name Fālibulā is coming from Philippa or Kāynā from Cain, as in the Bible. But this mere speculation and I do not want to put any english or greek names into the record, just yet alone the latin translitteration from Dillmanns catalogue and the diverging Gəʿəz names. I don't know, how to refer to the Bodleian manuscript. There is no number in Dillmanns catalogue except for his own (XXII), but I don't find BDLaeth with the number 22.

* Update PRS14034FalibulaKayna

I corrected the names and the added the relation to BDLbruce82.
Yet alone the ID PRS14034FalibulaKayna remains in the garbled form - sorry for that!

* Update PRS14031John

I reduced the certainty of the birth date, as requested by Eugenia.

* Updated PRS7054Michael

I updated names and relations in PRS7054Michael.

* Update PRS7054Michael

Updated faith, occupation and floruit of PRS7054Michael.

* Update PRS7054Michael

I don't know, how to relate PRS7054Michael to LOC7371Atrib, that has no paths-ID (as far as I have figured out). LOC-ID or Wikidata do not work or need another way of encoding.

* Update PRS7054Michael

Updated Aṯrīb and Malīǧ