Updated floruits of PRS records in Tanasee1 (no. 9)
I added or changed a floruit in several PRS records of persons, that are mentioned in two different additions on 85va and 232ra. This is possible, because I think, that one person (PRS14129Damye) is mentioned in both places and can be traced to the 17th/first half of 18th century.