New PRS records for Tanasee1 (no. 17)
I created three more records for persons, that I found in Tanasee1 on folio 232va. I calculate this addition to the 17th or first half of 18th century according to the palaeography.
I am not certain for PRS14205Marqoryos. I think, he is a person, who donated some manuscripts as well, but it may perhaps also be a title of a book, that was donated. As well, I am not certain about the title falāsi. I think it refers to a monk, who secluded himself and not to a member of Beta ʾIsrāʾel. I case, it is monk, is it appropriate to encode falāsi as an occupation or rather a title?