Create PRS14209Daneel in Tanasee1 (no. 19)
I found a person with the name Dānəʾəl in an addition in Tanasee1 on folio 234r.
He bears the title ʾabbā and the title ʾabbāmakān. He shall be contemporary of PRS8595SayfaAr and PRS8387SalamaI. However, I think the note is much older because of the palaeography and the position on two folia, that I think belong to different quires, that have been joined in later time.