Update 3 PRS records in Tanasee1 (no. 17)
I found the name Tomās in another addition on folio 234vb, that I like to date very early to the second half of the 14th or first half of the 15th century. It is probably the same Tomās, that is also mentioned here on 232va. For this reason, I deleted the floruit, because I think it is too vague to make a judgement on the dating of the addition based on palaeography.