Update and create PRS IDs in Tanasee1 (no. 20)
I found two new PRS IDs in Tanasee1 on folio 234va. I dated the floruits of these two persons and the already existing PRS14086Qalementos very early to the second half of the 14th or the first half of the 15th century because of the name of king ዳዊት mentioned above. I am puzzled as to which king might be mentioned here and have decided on David I as the earliest possible, but others are also possible. How should I proceed if the dating is not certain? Should I leave ዳዊት unrecorded and delete the floruit? Or can this addition be dated by deducing the palaeographic forms?