Persons in BMLacq384
I created a PRS ID for the owner Walatta ʾIyāsus from the 19th/20th century. There are plenty of persons with this name already, but it is impossible to identify the one or the other with the one mentioned in this manuscript.
I updated the record for Mary Magdalene and created a new one for Sara. Both are expected to be followers of Jesus as we discussed in issue #2440.