Created PRS for BMLacq679
I created four PRS IDs for persons, who are mentioned in BMLacq679 as discussed in #2310.
I came to the conclusion, that the person PRS14220HeggaYohannes is the father of PRS14221Menilek. I am still puzzled, whether it is possible to establish a family relation also to the others.
PRS14222TawaldaMadhen may be seen, that she/he is inheritor or inheritant of the former, if we accept this name connected to አርስት (succession, inheritance, etc.).
PRS14223Maqaryos is mentioned as ራብአዊ , but it is not clear, what ራብአዊ is standing for. He may be the fourth in the line of inheritance (e.g. the great grandson) or the fourth in this list or blessed people. It is rather unlikely, that he is a high ranking dignitary as a patriarch or abbot, because he has not title, that could indicate that.