PRS in Tanasee1 (no.15) (#759)
* PRS in Tanasee1 (no.15)

I found more personal names in another addition. The height of the characters is rather small and for this reason some are barely legible. This is especially true for the glosses, that are written next to this addition.

* Updated floruits of PRS in Tanasee1 (no. 15)

I updated the floruits in a number of PRS records that appear together in an addition on f. 24rb. I came to the conclusion, that it is appropriate to give a more precise floruit, because one of them (PRS14194Atnatewos) is most likely the same person, who also shows up in another addition on folio 232ra, that might be dated to the 17th or first half of 18th centuries.

* Updated PRS in Tanasee1 (no. 15)

I added ፡ to names according to Eugenias request.