Persons in Tanasee1 (no. 18) (#762)
* Persons in Tanas)ee1 (no. 18

I found some persons in Tanasee1 on folio 233rb. I identified ንጉሥ፡ ፀሐይ፡ ሰገድ፡ as PRS10473Yostos and therefore defined the floruit of the three new PRS records to the end of 17th century and first half of 18th century.
I was puzzled about the identity of the person named ʾIyasu, because I suspected him possibly to be the earlier or of the later kings with the same name, but found this unrealistic from the context and encoded him as a new individual. He might have been a local dignitary, who played a role in this transaction.

* Update PRS ID in Tanasee1 (no. 18)

I changed all ə to ǝ according to Eugenias request.