PRS in leporello from Luzern (#769)
* PRS in leporello from Luzern

I have created three PRS IDs for persons in the leporello from Luzern, that has the preliminary shelfmark MsEth1.
Silvia and Walter Frei can be found on diverse webpages, that exhibit the dates of birth and the date of decease for Walter Frei.
For PRS14226WaldaIyasus, I cannot say whether his name is just Walda ʾIyāsus or whether the name includes also the following one or two words (ከበደ፡ ይርሳው). I included Kabbada to the name, because this is familiar to me. I do not know, if yərsāw is also a name or an epiteth or a phrase.

* Update PRS14226WaldaIyasus

I updated the name of Walda ʾIyāsus Kabbada Yərsāw.

* Update PRS14226WaldaIyasus

I changed all ə to ǝ, as Eugenia has requested it.

* Updated PRS records in leporello

I updated the records according to Eugenias and Denis' requests.

* Update PRS14225WalterFrei

I updated the dates of birth and death according to Eugenias request.

* Update records Walter and Silvia Frei

I inserted a space to the name according to Eugenias request.