New PRS records for Tanasee1 (no. 17) (#761)
* New PRS records for Tanasee1 (no. 17)

I created three more records for persons, that I found in Tanasee1 on folio 232va. I calculate this addition to the 17th or first half of 18th century according to the palaeography.
I am not certain for PRS14205Marqoryos. I think, he is a person, who donated some manuscripts as well, but it may perhaps also be a title of a book, that was donated. As well, I am not certain about the title falāsi. I think it refers to a monk, who secluded himself and not to a member of Beta ʾIsrāʾel. I case, it is monk, is it appropriate to encode falāsi as an occupation or rather a title?

* Update 3 PRS records in Tanasee1 (no. 17)

I found the name Tomās in another addition on folio 234vb, that I like to date very early to the second half of the 14th or first half of the 15th century. It is probably the same Tomās, that is also mentioned here on 232va. For this reason, I deleted the floruit, because I think it is too vague to make a judgement on the dating of the addition based on palaeography.

* Update PRS records in Tanasee1

I corrected and updated some records from Tanasee1 (no. 17).

* Updated PRS records from Tanasee1 (no. 17)

I changed ə to ǝ and changed the titles according to Eugenias requests.

* Validate PR

I deleted the disturbing phrases and conflicts.