Prs in tanasee1  - no. 13 (#757)
* New PRS in Tanasee1 (no. 13)

I found another addition in Tanasee1 on folio. 24ra. I am tempted to date this addition and the names included to the 17th century because of the palaeography and because of the position in the codex. However, I want to be more cautious in giving a date and encoded every person as 16th to 18th century.
About PRS14186Pawlos, I am not sure, whether the title ʾabuna may point to the Ethiopian bishop or to a local church leader from Kəbrān and if I could encode him as such. However, I did not find any record with the name Ṗawlos as an (arch-)bishop.

* New PRS in Tanasee1 (no. 13)

I found another addition in Tanasee1 on folio. 24ra. I am tempted to date this addition and the names included to the 17th century because of the palaeography and because of the position in the codex. However, I want to be more cautious in giving a date and encoded every person as 16th to 18th century. It would be nice, if you can have a look on the images and tell me, if you agree with the dating 16th to 18th century or whether you recommend to be more precise on that.

About PRS14186Pawlos, I am not sure, whether the title ʾabuna may point to the Ethiopian bishop or to a local church leader from Kəbrān and if I could encode him as such. However, I did not find any record with the name Ṗawlos as an (arch-)bishop.

* Update PRS14186Pawlos

I updated PRS14186Pawlos according to Eugenias request.

* Updated PRS records in Tanasee1 (no. 13)

I added ፡ to names and changed all ə to ǝ according to Eugenias requests.