Persons in BMLacq384 (#766)
* Persons in BMLacq384

I created a PRS ID for the owner Walatta ʾIyāsus from the 19th/20th century. There are plenty of persons with this name already, but it is impossible to identify the one or the other with the one mentioned in this manuscript.
I updated the record for Mary Magdalene and created a new one for Sara. Both are expected to be followers of Jesus as we discussed in issue #2440.

* Update PRS14218Sara

I added the Wikidata Q238261 for Sara to be identified with Saint Sarah, although this is not entirely certain according to Eugenias post in #2440.

* Updated PRS record in BMLacq384

I updated the records according to Eugenias requests.

* Update floruit PRS14218Sara

I updated PRS14218Sara according to Eugenias request.

* Update PRS14218Sara

I updated the record according to Eugenias request.

* Updated PRS-records in BMLacq384

I updated the records according to Dorotheas request.