Create PRS14230GBRaimondi (#779)
* Create PRS14230GBRaimondi

I created an ID for Giovanni Battista Raimondi. He claimed by Assemani to be the author of the lexicon in BMLor321. The endeavours of this person is described in bm:Farina2018RaimondisTravel, however I think it is not our policy to refer to secondary literature in PRS records, correct?

* Update PRS14230GBRaimondi

I updated the record according to Eugenias request.

* Update PRS14230GBRaimondi

I updated the record according to Eugenias request.