Create PRS for BLorient13223
I created three records for scribes recorded in BLorient13223 and for one owner of this manuscript. For PRS14346WaldaAb Strelcyn recorded the form Wald ʾAb in the manuscript, but reasonably emended that name to Walda ʾAb. I am not sure, if this second name and this explanation should be given in the record or not.