PRS in BLorient2083
I want to create six PRS records for persons, that are attested in BLorient2083. PRS14423Malakotawit, who commissioned the manuscript might be identical with a prominent figure. However this is not certain but rather speculation. Do you agree to mark this possible identification with dc:relation or do you think this is too speculative?