PRS in BLorient8818
I created PRS IDs for owners of BLorient8818. The name of Mr. Lieder is present in a letter attached to the manuscript as accompanying material. Stephen Emmel explained in his contribution, that B. R. T. is a misreading for J. R. T. (Lieder). I took the opinion of Emmel and marked the dispute in an abstract. I identified J. R. T. Lieder as a reverend and husband of Egyptologist Alice Lieder and created also a record for her.