Update PRS records according to Nafisas remark
The reason for determining the floruit to 1894 was the acquisition of the manuscript by the British Museum in that year. However, I think it is not neccessarily the case, that people die, when their manuscripts are entering a library. I think, the floruits are in most case rather estimations and for that reason, I decided to change ist to `<floruit notBefore="1700" notAfter="1899">18th/19th century</floruit>` (both in PRS14553WalattaMaryam and PRS14552WKTaytu).