Owners of BLorient12959
I did not find any person with the name J. Malcolm-Davies or ʾƎssǝta Śǝllāse.

However, when I create a record for ʾƎssǝta Śǝllāse, I did the mistake to type EssetaSǝllāse to the ID in the form and this resulted to a mistake because of the letter ǝ, that is not allowed in an ID. I could not change the ID and for this reason, I created a new one, which by mistake ended up with PRS14563EssetaSellase_ (with a downspace '_' in the end). I hope, that this downspace is not a big problem. Otherwise, I possibly need to create again a new ID. I am sorry for this turmoil!