Updated ETH records on Galla etc.
I wanted to add some information in an abstract, but I have seen, that there is no option to add an abstract to any PRS or ETH record. It would be possible to add information in <occupation> or <floruit>, but this seemed to be not suitable for a discussion of the term, which I intended and outlined in https://github.com/BetaMasaheft/Documentation/issues/2637. For this reason, I refrained from that and limited myself with small updates of the bibliographies and transcriptions in each of the records.