Create PRS in BLorient12860 (#873)
* Create PRS in BLorient12860

I created record for two persons, that are mentioned in BLorient12860 - both most probably as owners. For Marius Bernard, I assume that it is the author of various orientalist books (wd:Q110089719). However, he is mentioned in the manuscript with his title "Angelus Comnenus", but I was not able to verify this title for Q110089719.

For Т. Печагуев, I did not find any information. I assume, that it is a Russian name, but it can also be Serbian, Bulgarian or Ukrainian.

* Update PRS14477Pechaguev according to Denis requests