Owners of BLorient11804
I tried to find a person with the name 'C. Paine', but I not find either a 'Captain Paine' (or 'Captain Payne') nor any prominent person with a first name beginning with C. Therefore, I do not know, whether 'C.' refers to a title or to a first name. I expressed this insecurity in a short description in roleName.