Delete Zaffaru according to Eugenias suggestion
Eugenia convinced me, that Zaffaru can not refer to a person with the name Zaffara Śǝllāse, because the latter name śǝllāse is  a feminin form and cannot be referred to with -u. From this perspective it is likely, that Zaffaru refers to Zaffara Mikāʾel as suggested by Eugenia.