PRS in BNFabb68 (#1013)
* PRS in BNFabb68

I created PRS IDs, which are attested in BNFabb68.

* Update Zaffaru et alia

* Delete Zaffaru according to Eugenias suggestion

Eugenia convinced me, that Zaffaru can not refer to a person with the name Zaffara Śǝllāse, because the latter name śǝllāse is  a feminin form and cannot be referred to with -u. From this perspective it is likely, that Zaffaru refers to Zaffara Mikāʾel as suggested by Eugenia.

* Corrections according to Nafisas requests

* Create PRS14697Zaffaru.xml

* Update PRS14694ZafaraSel.xml

* Update PRS14698ZaffaraMik.xml

---------

Co-authored-by: Eugenia <eugenia.sokolinski@uni-hamburg.de>