Create PRS14718KenfaM
I created a new record for Kənfa Mikāʾel as he is attested in BMLacq681 according to your request in your email from 2025-03-06. I wanted to update also the relation in PRS13971Abagya, but I found each relation in the PRS record and in the ms record correct. So I do not know, what was the problem and if it is cured by now. I made no more than minor changes to this record.