Minor updates
I made some minor updates to three records:

Titles in PRS10393Yohanne, because I think we used to write all titles and occupations in PRS records non-capitalized.

I wrote the EMML nos. in the tags, because otherwise the manuscripts are not properly displayed.

I deleted addName in PRS9430Tewodros, because otherwise 'II' does not appear in the title and in the search results.