Up bishops in BLorient12913J
I updated the records of two bishops of the 20th century.

I was confused about the transcription of ጵ . Personally, I prefer ̣Pṗ, but here I found Ṗ (ṗ), but in the EAe, there is ̣P ̣̣p ̣. Do we have any straight rule? The Guidelines do not give an answer to this question in the Transliteration Principles.