Up PRS BLorient13308
I updated two records of personal names, which appear in BLorient13308.

PRS4359Gabrase hold important office in the time of Menilek II. He appears with the title አለቃ፡ in BLorient13308 as well as in IVEf9415, IVEf957 and EMML517. For the deletion of PRS12393GabraS (attested in RNBspbdaBII60) see https://github.com/BetaMasaheft/Documentation/issues/3147.

PRS11674WaldaIy appears as a scribe of a book ordered by Menilek II in BLorient13308 as well as in BAVet95. Without consulting the images of the two manuscripts, I assume, that it may be the same person to copied these two manuscripts in 1894/5 and 1914.