PRS (scribe) for BLorient13309
I created record for PRS15043Fanta. He seems to have been active in or around Gondar in the 17th or 18th century. He is possibly the same as PRS14381Fanta, who is attested in Tegray. However, as long as I have no access to the manuscript images, I would not dare to establish an identification. I wanted to mention this possible identification in a relation, but I did not know which is the correct relation for it.