PRS in ANLcr76
A person who wrote tigrinyā poems and a person, for whom a text of the Miracles of Mary was copied.

For PRS15077TaklaWM, his name was given with Taḵla Waldamāryām in the catalogue. I preferred to keep, what I think is the Tigrinya version of that name and added for convenience a conventional Geez-version. I quoted the title fitawrar in Tigrinya as well accompanied by the better known Geez-variant of this title. For šum bāšā, I was not certain, whether it is a title or an office. EAe has šum bāš as a title. So I preferred to define it as title in type.