PRS in ANLcr133 (owners)
I did not add a floruit in each case, because the time span of more than 300 years is possible and shall not be encoded according to our usual practice.

I did not use snap:FatherOf to mark the relation between Ṣamra Qāl and Ləbsa Qāl, because this would lead to reciprocal relations, which we want to avoid in future.

I was not certain about the name Ṣamra Qāl, which is possibly Śamra Qāl, but I did not dare to change the name, which is provided by Stefan Strelcyn in his catalogue.