Create PRS15175DeRossi.xml
I do not know, which relation is best to mark a PRS record of a person, who is possibly the same as another person. I used dc:relation in the past, but I learnt, that this is not the suitable relation. Is skos:exactMatch with a desc Ok for it?